pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sherlock"
version = "1.0.0"
description = "Sherlock — Intelligent New Relic investigation MCP server"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "NewRelic MCP Team" },
]
keywords = ["newrelic", "mcp", "model-context-protocol", "observability", "incident-response"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "tenacity>=8.0.0",
    "keyring>=24.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
sherlock = "main:main_sync"

[tool.hatch.build.targets.wheel]
packages = ["client", "core", "tools", "scripts"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
target-version = "py311"

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
